agent_security_navigator/requirements.txt
fastapi>=0.110
uvicorn[standard]>=0.23
python-dotenv>=1.0
pydantic>=2.4
requests>=2.31
faiss-cpu>=1.7.4
sentence-transformers>=2.7.0
pypdf>=4.2.0
numpy>=1.26
